pyproject.toml
[project]
name = "macocr"
authors = [
    {name = "Rio Fujita", email = "rifujita@microsoft.com"},
]
version = "0.1.0"
license = {file = "LICENSE"}
description = "macocr is a python script for OCR on macOS"
readme = "README.md"

requires-python = ">=3.11"
dependencies = [
    "ocrmac>=1.0.0",
    "pillow-heif>=0.21.0",
    "python-magic>=0.4.27",
]

[project.urls]
Homepage = "https://github.com/rioriost/homebrew-macocr"
Issues = "https://github.com/rioriost/homebrew-macocr/issues"

[project.scripts]
macocr = "macocr.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/macocr"]

[tool.hatch.build.targets.sdist]
include = [
  "src/macocr/*.py",
  "*.py",
  "images/*",
  "results/*",
]
exclude = [
  "macocr.rb",
  "uv.lock",
  "dist/.DS_Store",
]
